Factor out kafka async wrapper code
This code has been moved to a separate module:
  @wikimedia/node-rdkafka-factory

and added here as an optional dependency.  The default-eventgate
factory will only attempt to require this if kafka.conf is set.

By factoring out this code, custom EventGate implementations
can require is separately from EventGate to create their own
Kafka produce functions.

Bug: T266058
