+++ .gitignore
@@ -8,4 +8,4 @@ npm-debug.log
 package-lock.json
 .eslintcache
 .nyc_output
-
+output.json
