--- config.dev.yaml
@@ -1,49 +1,0 @@
-# EventGate example config file.
-#
-# Configures the service-runner as well as the EventGate application.
-# See: https://github.com/wikimedia/service-runner#config-loading
-
-# Number of worker processes to spawn.
-# Set to 0 to run everything in a single process without clustering.
-# Use 'ncpu' to run as many workers as there are CPU units
-num_workers: 0
-
-# Log error messages and gracefully restart a worker if v8 reports that it
-# uses more heap (note: not RSS) than this many mb.
-worker_heap_limit_mb: 200
-
-# Logger info
-logging:
-  level: info
-
-services:
-  - name: eventgate-dev
-    # a relative path or the name of an npm package, if different from name
-    module: ./app.js
-    # optionally, a version constraint of the npm package
-    # version: ^0.4.0
-    # per-service config
-    conf:
-      port: 8192
-      # Events can be large; increase max body size
-      max_body_size: 4mb
-
-      # more per-service config settings
-      user_agent: eventgate-dev
-
-      eventgate_factory_module: '../lib/factories/dev-eventgate'
-
-      # This field in each event will be used to extract a
-      # (possibly relative) schema uri.  The default is $schema.
-      # An array of field names will cause EventGate to search for
-      # fields by these names in each event, using the first match.
-      schema_uri_field: $schema
-
-      # If set, these URIs will be prepended to any relative schema URI
-      # extracted from each event's schema_field.  The resulting URLs will
-      # be searched until a schema is found.  Change this
-      # to match paths to your local schema repositories.
-      schema_base_uris: [./test/schemas]
-
-    #   output_path: ./output.json
-
