+++ lib/factories/default-eventgate.js
@@ -17,6 +17,14 @@ const {
  * to make new functions that extract information from events (e.g. schema_uri_field)
  * and to create validate and produce functions for constructing the EventGate instance.
  *
+ * If kafka.conf is provided, node-rdkafka-factory (and node-rdkafka) will be
+ * required and the this EventGate will produce to Kafka.  node-rdkafka-factory
+ * is an optional dependency, so make sure it is properly installed.
+ *
+ * If output_path is configured, this EventGate will produce to output_path.
+ * Both kafka and file output are simulataneously supported, depending
+ * on what is configured.
+ *
  * You can and should make new EventGate factory modules that themselves return
  * an custom EventGate factory.  This module also exports the individual functions
  * used to create this default EventGate, in case you want to plug some of them
@@ -51,8 +59,14 @@ const {
  *      Their stream fields at different locations.
  *      Default: undefined
  *
+ * - output_path
+ *      If set, valid events will be written to this file.
+ *      A value of 'stdout' will write events to stdout.
+ *      Default: stdout.
+ *
  * - kafka.conf
  *      node-rdkafka KafkaProducer configuration
+ *      If set, events will be produced to Kafka.
  *
  * - kafka.topic_conf
  *      node-rdkafka KafkaProducer topic configuration
@@ -64,12 +78,14 @@ const defaultOptions = {
     schema_base_uri:        undefined,
     schema_file_extension:  undefined,
     stream_field:           undefined,
-    kafka: {
-        conf: {
-            'metadata.broker.list': 'localhost:9092'
-        },
-        topic_conf: {}
-    },
+    output_path:            'stdout',
+    // disable Kafka produce by default so we can have node-rdkafka as an optional dependnecy.
+    // kafka: {
+    //     conf: {
+    //         'metadata.broker.list': 'localhost:9092'
+    //     },
+    //     topic_conf: {}
+    // },
 };
 
 /**
@@ -131,7 +147,6 @@ function makeExtractStream(options) {
     }
 }
 
-
 /**
  * Creates a new schema URI based validate(event) function.
  * @param {Object} options
@@ -168,10 +183,92 @@ function makeValidate(options, logger) {
 }
 
 /**
- * Creates a function that returns a connected Kafka Producer's produce function,
- * suitable for passing to EventGate as the produce function argument.
+ * Creates a function that writes events to output_path.
+ * @param {Object} options
+ * @param {Object} options.output_path
+ * @param {Object} logger
+ * @param {Object} metrics
+ * @return {EventGate~produce} (event, context) => Promise<event>
+ */
+function makeFileProduce(options, logger, metrics) {
+    const fs = require('fs');
+    const {
+        once
+    } = require('events');
+
+    const writeOptions = { flags: 'as' };
+    if (options.output_path === 'stdout') {
+        // If fd is set, createWriteStream will ignore output_path.
+        writeOptions.fd = process.stdout.fd;
+    }
+    const outputStream = fs.createWriteStream(options.output_path || './output.json', writeOptions);
+
+    return async (event, context = {}) => {
+        const serializedEvent = Buffer.from(JSON.stringify(event) + '\n');
+        if (!outputStream.write(serializedEvent)) {
+            await once(outputStream, 'drain');
+        }
+    };
+}
+
+/**
+ * Creates a function that returns function that produces events to Kafka.
  * NOTE: This function uses undefined Kafka key and partition when producing.
  * If you need to set key and/or partition, you should make your own produce function.
+ * @param {Object} options
+ * @param {string} options.stream_field
+ *      Used to construct the kafka topic
+* @param {string} options.schema_uri_field
+ *      Used to construct the kafka topic if stream_field is not set.
+ * @param {Object} options.kafka.conf
+ *      node-rdkafka KafkaProducer configuration
+ * @param {Object} options.kafka.topic_conf
+ *      node-rdkafka KafkaProducer topic configuration
+ * @param {Object} logger
+ * @param {Object} metrics
+ *      service-runner metrics object, used if statistics.interval.ms is set
+ *      to collect rdkafka metrics.
+ * @return {Promise<{EventGate~produce}>} (event, context) => Promise<event>>
+ */
+async function makeKafkaProduce(options, logger, metrics) {
+    const kafka = require('node-rdkafka-factory');
+
+    // This default EventGate implementation uses the stream_name
+    // extracted from the event as the topic.
+    const extractTopic = makeExtractStream(options);
+
+    const kafkaProducer = await kafka.producerFactory(
+        'guaranteed',
+        options.kafka.conf,
+        options.kafka.topic_conf
+    );
+
+    // If rdkafka is configured with statistics reporting turned on,
+    // AND if we are given a metrics object (node-statsd like or from service-runner app directly)
+    // then enable Kafka metrics reporting using node-rdkafka-statsd.
+    if (options.kafka.conf['statistics.interval.ms'] && metrics) {
+        const rdkafkaStatsd  = require('node-rdkafka-statsd');
+        logger.info('Enabling Kafka metrics reporting for Kafka producer');
+
+        kafkaProducer.on(
+            'event.stats',
+            rdkafkaStatsd(metrics.makeChild('rdkafka.producer'))
+        );
+    }
+
+    // Return a new function that takes a single event argument for produce.
+    return (event, context = {}) => {
+        const topic = extractTopic(event);
+        const serializedEvent = Buffer.from(JSON.stringify(event));
+        return kafkaProducer.produce(topic, undefined, serializedEvent, undefined);
+    };
+}
+
+
+/**
+ * Creates a function that will write events to an output file and/or
+ * produce events to Kafka.
+ *
  * @param {Object} options
  * @param {string} options.stream_field
  *      Used to construct the kafka topic
@@ -182,20 +279,39 @@ function makeValidate(options, logger) {
  * @param {string} options.partition_field
  *      If given, the event's 'partition' will be extracted from the event at this
  *      dotted object path.
- * @param {KafkaProducer} producer
- *      Connected node-rdkafka Producer.
- * @return {EventGate~produce} (event, context) => Promise<event>
+ * @param {Object} options.output_path
+ *      If given, events will be written to this file.  If 'stdout' events will be
+ *      written to stdout.
+ * @param {Object} options.kafka.conf
+ *      If given, events will be produced to Kafka.
+ *      node-rdkafka KafkaProducer configuration
+ * @param {Object} options.kafka.topic_conf
+ *      node-rdkafka KafkaProducer topic configuration
+ * @param {Object} logger
+ * @param {Object} metrics
+ *      service-runner metrics object.
+ * @return {Promise<{EventGate~produce}>} (event, context) => Promise<event>>}
  */
-function makeProduce(options, producer) {
-    // This default EventGate implementation uses the stream_name
-    // extracted from the event as the topic.
-    const extractTopic = makeExtractStream(options);
+async function makeProduce(options, logger, metrics) {
+
+    let fileProduce;
+    if (options.output_path) {
+        fileProduce = makeFileProduce(options, logger, metrics);
+        logger.info('Writing valid events to ' + options.output_path);
+    }
+
+    let kafkaProduce;
+    if (options.kafka && options.kafka.conf) {
+        kafkaProduce = await makeKafkaProduce(options, logger, metrics);
+        logger.info('Producing valid events to Kafka at ' + options.kafka.conf['metadata.broker.list']);
+    }
 
     // Return a new function that takes a single event argument for produce.
-    return (event, context = {}) => {
-        const topic = extractTopic(event);
-        const serializedEvent = Buffer.from(JSON.stringify(event));
-        return producer.produce(topic, undefined, serializedEvent, undefined);
+    return async (event, context = {}) => {
+        if (fileProduce) {
+            await fileProduce(event, context);
+            await kafkaProduce(event, context);
+        }
     };
 }
 
@@ -214,9 +330,9 @@ function makeProduce(options, producer) {
  * @param {string} options.stream_field
  *      Optionally used to extract the event's destination stream name.
  * @param {Object} options.kafka
- * @param {Object} options.kafka.options
+ * @param {Object} options.kafka.conf
  *      node-rdkafka KafkaProducer configuration
- * @param {Object} options.kafka.topic_options
+ * @param {Object} options.kafka.topic_conf
  *      node-rdkafka KafkaProducer topic configuration
  * @param {Object} logger
  * @param {Object} metrics service-runner statsd-like metrics interface.  This is provided from
@@ -224,46 +340,15 @@ function makeProduce(options, producer) {
  * @return {Promise<EventGate>}
  */
 async function eventGateFactory(options, logger, metrics) {
-    const kafka          = require('../kafka');
-
     // Set default options
     _.defaults(options, defaultOptions);
-
-    const kafkaProducer = await kafka.producerFactory(
-        'guaranteed',
-        options.kafka.conf,
-        options.kafka.topic_conf
-    );
-
-    // If we Kafka is set with statistics reporting turned on,
-    // AND if we are given a metrics (node-statsd like or from service-runner app directly)
-    // then enable Kafka metrics reporting using node-rdkafka-statsd.
-    if (options.kafka.conf['statistics.interval.ms'] && metrics) {
-        let rdkafkaStatsd;
-        try {
-            rdkafkaStatsd  = require('node-rdkafka-statsd');
-        } catch (err) {
-            logger.fatal(
-                err,
-                'node-rdkafka-statsd could not be required. Make sure it is installed properly.'
-            );
-            process.exit(1);
-        }
-
-        logger.info('Enabling Kafka metrics reporting for Kafka producer');
-
-        kafkaProducer.on(
-            'event.stats',
-            rdkafkaStatsd(metrics.makeChild('rdkafka.producer'))
-        );
-    }
 
     return new EventGate({
         // This EventGate instance will use the EventValidator's
         // validate function to validate incoming events.
         validate:   makeValidate(options, logger),
         // This EventGate instance will use a kafka producer
-        produce:    makeProduce(options, kafkaProducer),
+        produce:    await makeProduce(options, logger, metrics),
         log:        logger
     });
 }
@@ -274,5 +359,7 @@ module.exports = {
     makeExtractStream,
     makeExtractSchemaUri,
     makeValidate,
-    makeProduce
+    makeProduce,
+    makeFileProduce,
+    makeKafkaProduce
 };
