+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "eventgate",
-  "version": "1.4.1",
+  "version": "1.5.0",
   "description": "Event intake service - POST JSONSchemaed events, validate, and produce.",
   "main": "./index.js",
   "scripts": {
-    "start": "./server.js -c ./config.dev.yaml",
+    "start": "./server.js -c ./config.yaml",
     "test": "PREQ_CONNECT_TIMEOUT=15 nyc mocha && npm run lint",
     "lint": "eslint --cache --max-warnings 0 --ext .js --ext .json lib routes app.js",
     "docker-start": "service-runner docker-start",
@@ -53,7 +53,7 @@
     "swagger-ui": "git+https://github.com/wikimedia/swagger-ui#master"
   },
   "optionalDependencies": {
-    "node-rdkafka": "~2.9.1",
+    "@wikimedia/node-rdkafka-factory": "^1.0.0",
     "node-rdkafka-statsd": "^0.1.2"
   },
   "devDependencies": {
