+++ test/schemas/test/0.0.2
@@ -42,3 +42,9 @@ properties:
     type: object
     additionalProperties:
       type: string
+examples:
+  - $schema: /test/0.0.2
+    meta:
+      stream: test
+      id: 12345678-1234-5678-1234-567812345678
+    test: example string
